--- apps/pesayetu/src/components/HURUmap/IndicatorTitle/Download.js
@@ -1,236 +1,0 @@
-import { ButtonBase, IconButton, Grid, Typography } from "@mui/material";
-import { useTheme } from "@mui/material/styles";
-import clsx from "clsx";
-import Image from "next/image";
-import PropTypes from "prop-types";
-import React, { useState, useEffect } from "react";
-import * as vega from "vega";
-
-import useStyles from "./useStyles";
-import {
-  downloadSheetData,
-  downloadJson,
-  downloadImage,
-  createImage,
-} from "./utils";
-
-import cfalogo from "@/pesayetu/assets/logos/Group4462.svg";
-import projectlogo from "@/pesayetu/assets/logos/Group5002.svg";
-import { idify } from "@/pesayetu/components/HURUmap/Chart/utils";
-import config, { hurumapArgs } from "@/pesayetu/config";
-
-function Download({
-  title,
-  chartValue,
-  handleChartValueChange,
-  disableToggle,
-  spec,
-  source,
-  height,
-  data,
-  currentFilters,
-  profileNames,
-  isCompare,
-  ...props
-}) {
-  const classes = useStyles(props);
-  const [view, setView] = useState(null);
-  const { palette } = useTheme();
-
-  const {
-    indicatorTitle: {
-      download: { values, layouts, imageTypes, fileTypes },
-    },
-  } = hurumapArgs;
-  const [layout, setLayout] = useState(0);
-
-  useEffect(() => {
-    try {
-      const viewProp = new vega.View(vega.parse(spec), { renderer: "none" });
-      setView(viewProp);
-    } catch (error) {
-      console.error("Error creating view", error);
-    }
-  }, [spec]);
-
-  const setImageLayout = async (e, type) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setLayout(type);
-  };
-
-  const splitString = (str) => {
-    // eslint-disable-next-line prefer-regex-literals
-    const regex = new RegExp(/\S.{1,42}\S(?= |$)/, "g");
-    const chunks = str.match(regex);
-    return chunks;
-  };
-
-  currentFilters?.forEach(({ name, value }) => {
-    const filterName = idify(name);
-    view?.signal(`${filterName}Filter`, true);
-    view?.signal(`${filterName}FilterValue`, value);
-  });
-
-  const chartTitle = splitString(title)?.slice(0, 3);
-  const subtitle = currentFilters?.reduce((acc, cur) => {
-    return `${acc}${cur.name}: ${cur.value},`;
-  }, "");
-  const secondaryName = isCompare
-    ? ` vs ${profileNames?.secondary?.split("-")[0]}`
-    : "";
-  const chartSubtitle = `${subtitle} Location: ${profileNames?.primary}${secondaryName}`;
-
-  const handleImageDownload = async (e, type) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    const totalHeight = height + 300; // chartHeight + extra space for legends, logo + title;
-    const imgType = type.toLowerCase();
-
-    const url = await createImage({
-      view,
-      totalHeight,
-      chartTitle,
-      chartSubtitle,
-      source,
-      projectlogo,
-      cfalogo,
-      backgroundColor: palette.common.white,
-      layout,
-      imgType,
-      scaleFactor: config.images.scaleFactor,
-    });
-    downloadImage(url, title, imgType);
-  };
-
-  const handleDataDownload = async (e, type) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    const fileType = type.toLowerCase();
-    const fileName = `${title}.${fileType}`;
-
-    if (fileType === "json") {
-      downloadJson(data, fileName);
-    } else {
-      downloadSheetData(data, fileType, fileName, title);
-    }
-  };
-
-  return (
-    <Grid container className={classes.root}>
-      {!disableToggle && (
-        <>
-          <Grid item container xs={12} className={classes.row}>
-            {values.map((v) => (
-              <Grid
-                item
-                xs={6}
-                key={v}
-                className={clsx(classes.button, {
-                  [classes.activeButton]: chartValue === v,
-                })}
-              >
-                <ButtonBase
-                  className={classes.text}
-                  onClick={() => handleChartValueChange(v)}
-                >
-                  {v}
-                </ButtonBase>
-              </Grid>
-            ))}
-          </Grid>
-          <Grid item xs={12} className={clsx(classes.row, classes.header)}>
-            <Typography className={classes.text}>Download chart as:</Typography>
-          </Grid>
-        </>
-      )}
-      <Grid item container className={classes.row}>
-        {imageTypes.map((p) => (
-          <Grid item xs={6} key={p} className={classes.button}>
-            <ButtonBase
-              className={classes.text}
-              onClick={(e) => handleImageDownload(e, p)}
-            >
-              {p}
-            </ButtonBase>
-          </Grid>
-        ))}
-      </Grid>
-      <Grid item xs={12} className={clsx(classes.row, classes.layout)}>
-        <Typography className={classes.text}>Layout option:</Typography>
-      </Grid>
-      <Grid item container className={classes.row}>
-        {layouts.map((p, index) => (
-          <Grid
-            item
-            xs={6}
-            key={p}
-            className={clsx(classes.button, {
-              [classes.activeButton]: layout === index,
-            })}
-          >
-            <IconButton
-              className={classes.layoutButton}
-              onClick={(e) => {
-                setImageLayout(e, index);
-              }}
-              size="large"
-            >
-              <Image src={p} width={24} height={24} alt="layout" />
-            </IconButton>
-          </Grid>
-        ))}
-      </Grid>
-      <Grid item xs={12} className={clsx(classes.row, classes.header)}>
-        <Typography className={classes.text}>Download data as:</Typography>
-      </Grid>
-      <Grid item container className={classes.row}>
-        {fileTypes.map((f) => (
-          <Grid item xs={4} key={f} className={classes.button}>
-            <ButtonBase
-              className={classes.text}
-              onClick={(e) => handleDataDownload(e, f)}
-            >
-              {f}
-            </ButtonBase>
-          </Grid>
-        ))}
-      </Grid>
-    </Grid>
-  );
-}
-
-Download.propTypes = {
-  title: PropTypes.string,
-  spec: PropTypes.shape({}),
-  currentFilters: PropTypes.arrayOf(PropTypes.shape({})),
-  data: PropTypes.arrayOf(PropTypes.shape({})),
-  disableToggle: PropTypes.bool,
-  chartValue: PropTypes.oneOf(["Value", "Percentage"]),
-  handleChartValueChange: PropTypes.func,
-  height: PropTypes.number,
-  source: PropTypes.string,
-  profileNames: PropTypes.shape({
-    primary: PropTypes.string,
-    secondary: PropTypes.string,
-  }),
-  isCompare: PropTypes.bool,
-};
-
-Download.defaultProps = {
-  title: undefined,
-  spec: undefined,
-  currentFilters: undefined,
-  data: undefined,
-  disableToggle: false,
-  chartValue: undefined,
-  handleChartValueChange: undefined,
-  height: 450,
-  source: undefined,
-  profileNames: undefined,
-  isCompare: undefined,
-};
-
-export default Download;
