+++ packages/hurumap-core/package.json
@@ -61,12 +61,14 @@
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.14.0",
     "@mui/utils": "^5.12.3",
+    "papaparse": "^5.4.1",
     "leaflet": "^1.9.4",
     "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.1",
-    "react-share": "^5.1.0"
+    "react-share": "^5.1.0",
+    "xlsx": "^0.18.5"
   },
   "dependencies": {
     "prop-types": "^15.8.1"
