+++ packages/hurumap-core/src/Download/index.js
@@ -1,0 +1,3 @@
+import Download from "./Download";
+
+export default Download;
