+++ packages/hurumap-core/src/Download/utils.js
